+++ src/types.ts
@@ -1,0 +1,10 @@
+export type ListItem = {
+  isCreate: boolean,
+  url: string,
+  content: string,
+  timestamp: number,
+  time?: Date,
+  oo?: number,
+  xx?: number,
+  approved?: boolean,
+}