+++ src/utils.ts
@@ -1,5 +1,8 @@
 // @ts-ignore
 import { unsafeWindow } from "monkeyClient";
+import mitt from 'mitt';
+
+export const emitter = mitt();
 
 export const _window = unsafeWindow || window;